break out common e2e test functions into helpers
